// lib/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const Gender = {
  M: 'M',
  F: 'F'
} as const

export type Gender = (typeof Gender)[keyof typeof Gender]


export const Role = {
  admin: 'admin',
  staff: 'staff'
} as const

export type Role = (typeof Role)[keyof typeof Role]


export const Position = {
  professor: 'professor',
  department_head: 'department_head',
  secretary: 'secretary',
  hr: 'hr',
  academic: 'academic',
  finance: 'finance',
  support_director: 'support_director',
  vice_dean: 'vice_dean',
  dean: 'dean'
} as const

export type Position = (typeof Position)[keyof typeof Position]


export const ProcessType = {
  draft: 'draft',
  submitted: 'submitted',
  approved: 'approved',
  rejected: 'rejected'
} as const

export type ProcessType = (typeof ProcessType)[keyof typeof ProcessType]